requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
cloudinary==1.44.1
Django==5.1.15
django-cloudinary-storage==0.3.0
django-htmx==1.27.0
django-modeltranslation==0.20.2
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.12.0
